refactor(#133): 헤더 경로 변경
