+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import '../styles/globals.css';
 import { Inter } from 'next/font/google';
 import { ToastContainer } from 'react-toastify';
 import Footer from '@/app/_components/Footer';
-import Header from '@/components/Header/Header';
+import Header from '@/app/_components/Header/Header';
 import type { Metadata } from 'next';
 import 'react-toastify/dist/ReactToastify.css';
