+++ tsconfig.json
@@ -30,7 +30,7 @@
     "**/*.tsx",
     ".next/types/**/*.ts",
     "public/icons/svg.js",
-    "src/components/Header"
+    "src/app/_components/Header"
   ],
   "exclude": ["node_modules"]
 }
